+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.3.2
 
   * Optional `onInit` function - The `StoreConnector` and `StoreBuilder` Widgets now accept an `onInit` function that will be run the first time the Widget is created (using Store.initState under the hood). The `onInit` function takes the Store as the first parameter, and can be used to dispatch actions when your Widget is first starting up. This can be useful for data fetching.
-  * `rebuildOnNull` boolean option - `StoreConnector` now has an optional boolean `rebuildOnNull`. If your `converter` function produces a null value in response to a store `onChange` event, it will not rebuilt the Widget using the `builder` function. This can be useful for Widgets that need to display information that has been removed from the Store, but needs to be displayed as it animates off the screen.
+  * `ignoreChange` function - `StoreConnector` now takes an advanced usage / optional function `ignoreChange`. It will be run on every store `onChange` event to determine whether or not the `ViewModel` and `Widget` should be rebuilt. This can be useful in some edge cases, such as displaying information that has been deleted from the store while it is animating off screen. 
   * Documentation updates
 
 ## 0.3.1